CultOfQoL: Add shrine fuel, Knucklebones speed, warmth fix
- Add Rotburn (MAGMA_STONE) as shrine brazier fuel with configurable weight
- Fix vanilla bug: shrine warmth was defined but never applied to WarmthBar
- Add Knucklebones animation speed multiplier (1x-5x)
- Fix Magnet Range Multiplier slider not rounding to 0.25 increments
- Switch all mods to ConfigurationManagerEnhanced dependency
- Fix Rebirth "murdered" death text (now shows as old age)
- Update all documentation and changelogs

+++ Cult of the Lamb/CultOfQoL/Patches/Followers/FollowerLevelUpPatches.cs
@@ -53,7 +53,7 @@ public static class FollowerLevelUpPatches
         // Recalculate with actual XPLevel
         const float clampedFactor = (float)(1.0 + (10.0 - 1.0) / 5.0);
         var traitMultiplier = __result / clampedFactor;
-        var unclampedFactor = (float)(1.0 + ((double)__instance.XPLevel - 1.0) / 5.0);
+        var unclampedFactor = (float)(1.0 + (__instance.XPLevel - 1.0) / 5.0);
         __result = unclampedFactor * traitMultiplier;
     }
